Added support for specifying multiple labels for images in the image_data_layer.
This adds support for lines in the image_data_layer with the following format:

filename label1 label2 label3;ignore1 ignore2

Here multiple labels (label 1, 2, and 3) are associated with the data line and
labels to ignore (e.g. not take into account when calculating loss or accuracy)
can also be specified.

The separator for the two label lists can be specified by the user, as can the
separator for the individual labels.

By default, labels to be ignored are assigned a value of -1, but this can be
specified.

The SigmoidCrossEntropy layer has also been changed to allow for a particular
label to be ignored when calculating the loss, or back propagating the diff.